code_generator.py now only update outdated files in ./src/ast/     - `code_generator.py` do not update up-to-date files, so that they         are not considered as changed by CMake.     - clang-format is now called by `code_generator.py`     - clang-format on template files only for version >= 4

+++ src/language/code_generator.py
@@ -1,9 +1,24 @@
-import jinja2
+import argparse
+import filecmp
+import logging
+import os
 from pathlib import Path
+import shutil
+import subprocess
+import tempfile
+
+import jinja2
 
 from parser import LanguageParser
 import node_info
 import utils
+
+
+logging.basicConfig(level=logging.INFO, format='%(message)s')
+parser = argparse.ArgumentParser()
+parser.add_argument('--clang-format')
+args = parser.parse_args()
+clang_format = args.clang_format
 
 # parse nmodl definition file and get list of abstract nodes
 nodes = LanguageParser("nmodl.yaml").parse_file()
@@ -16,6 +31,8 @@ env = jinja2.Environment(loader=jinja2.FileSystemLoader(str(templates)),
                          lstrip_blocks=True)
 env.filters['snake_case'] = utils.to_snake_case
 
+updated_files = []
+
 for fn in templates.glob('*.[ch]pp'):
     if fn.name.startswith('py'):
         filename = basedir / 'pybind' / fn.name
@@ -24,5 +41,25 @@ for fn in templates.glob('*.[ch]pp'):
     else:
         filename = basedir / 'ast' / fn.name
     template = env.get_template(fn.name)
-    with filename.open('w') as fd:
-        fd.write(template.render(nodes=nodes, node_info=node_info))
+    content = template.render(nodes=nodes, node_info=node_info)
+    if filename.exists():
+        # render template in temporary file
+        # and update target file ONLY if different
+        # to save a lot of build time
+        fd, tmp_path = tempfile.mkstemp()
+        os.write(fd, content.encode('utf-8'))
+        os.close(fd)
+        if clang_format:
+            subprocess.check_call([clang_format, '-i', tmp_path])
+        if not filecmp.cmp(str(filename), tmp_path):
+            shutil.move(tmp_path, filename)
+            updated_files.append(str(fn.name))
+    else:
+        with filename.open('w') as fd:
+            fd.write(content)
+            updated_files.append(str(fn.name))
+        if clang_format:
+            subprocess.check_call([clang_format, '-i', filename])
+
+if updated_files:
+    logging.info('       Updating out of date template files : %s', ' '.join(updated_files))
